+++ src/types/puppeteer-chromium-resolver.d.ts
@@ -1,0 +1,22 @@
+declare module 'puppeteer-chromium-resolver' {
+    import { Browser } from 'puppeteer-core';
+    
+    interface PCRStats {
+        executablePath: string;
+        puppeteer: {
+            launch(options?: any): Promise<Browser>;
+        };
+        browser?: Browser;
+        browserWSEndpoint?: string;
+    }
+
+    const PCR: (options?: {
+        folderName?: string;
+        revision?: string;
+        installPath?: string;
+        proxy?: string;
+        host?: string;
+    }) => Promise<PCRStats>;
+
+    export default PCR;
+}