+++ webpack.config.js
@@ -1,0 +1,41 @@
+const path = require('path');
+
+module.exports = {
+  target: 'node',
+  mode: 'production',
+  entry: './src/extension.ts',
+  output: {
+    path: path.resolve(__dirname, 'out'),
+    filename: 'extension.js',
+    libraryTarget: 'commonjs2'
+  },
+  externals: {
+    vscode: 'commonjs vscode',
+    bufferutil: 'bufferutil',
+    'utf-8-validate': 'utf-8-validate'
+  },
+  resolve: {
+    extensions: ['.ts', '.js']
+  },
+  module: {
+    rules: [
+      {
+        test: /\.ts$/,
+        exclude: /node_modules/,
+        use: [
+          {
+            loader: 'ts-loader'
+          }
+        ]
+      }
+    ]
+  },
+  ignoreWarnings: [
+    {
+      module: /yargs|puppeteer-chromium-resolver/
+    }
+  ],
+  node: {
+    __dirname: false
+  }
+};