+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## [1.1.0] - 2024-11-21
+
+### Added
+- Alternative browser-based scraping method
+- Method selection dropdown in UI
+
 ## [1.0.3] - 2024-11-20
 
 ### Changed